move KeyExchange to separate module
